Changes for the Item classes, see #147 Fix IsVoid() check Make interface slots array protected, call interface[] instead

+++ Chraft/Interfaces/Containers/PersistentContainer.cs
@@ -42,7 +42,7 @@ namespace Chraft.Interfaces.Containers
         public WorldManager World;
         public UniversalCoords Coords;
         protected virtual ItemInventory[] Slots { get; set; }
-        public int SlotsCount;
+        public short SlotsCount;
         public virtual ItemInventory this[int slot]
         {
             get
